backend/requirements.txt
fastapi
uvicorn
sqlmodel
feedparser
pydantic-settings
requests
faiss-cpu
sentence-transformers
numpy
deep-translator
google-generativeai
python-jose[cryptography]
passlib[argon2]
python-multipart
argon2-cffi
python-dotenv



